Create separate file for `SortIcon`.
